app/requirements.txt
deepagents
pytest
langchain-openai
langchain-anthropic
fastapi==0.115.5
uvicorn[standard]==0.32.1
websockets==14.1
python-dotenv==1.0.1
pydantic==2.10.3
aiofiles==24.1.0
